+++ packages/icons/src/svg/index.js
@@ -5,6 +5,7 @@ import TriangleAlert from './triangle_alert.svg';
 import Close from './close.svg';
 import Visibility from './visibility.svg';
 import VisibilityOff from './visibility_off.svg';
+import ArrowDown from './arrow_down.svg';
 
 export {
   Star,
@@ -14,4 +15,5 @@ export {
   Close,
   Visibility,
   VisibilityOff,
+  ArrowDown,
 };