+++ packages/yoga/src/Theme/themes/BaseTheme.js
@@ -17,6 +17,7 @@ import {
   text,
   rating,
   tag,
+  dropdown,
 } from './components';
 
 const {
@@ -116,6 +117,15 @@ const BaseTheme = ({ primary, secondary, tertiary }) => {
       radii,
       fontSizes,
       fontWeights,
+    }),
+    dropdown: dropdown({
+      colors,
+      fontSizes,
+      fontWeights,
+      lineHeights,
+      radii,
+      spacing,
+      borders,
     }),
   };